Add files via upload
Touch talk Friday 9/15 coding session with Moulay and Andrew.

LCD Screens:
image.py is example code directly from the web
image_2screens.py is a single python file that displays two images (note that the reset wire cannot be shared by both screen to simultaneously display images so i disconnected the reset wire)
image_A.py was a test by andrew to see if we could display two images simulatenously with two separate files(run from terminal one at a time)

button_press_audioplay.py was an attempt to use the default 3.5mm port on the Rpi 3 to play music when a button input was detected.  Because the sound test code for the terminal runs forever it caused issues after we ran the code succesfully the first time.

To fix the button_press_audioplay code we need to figure out the audio player on the Rpi 3 and also play a simple test wave file.  Do this in terminal first, then do os.system.